+++ src/enums/starClassNamesEnum.ts
@@ -1,0 +1,4 @@
+export enum STAR_CLASS_NAMES {
+  FORM = 'rating-label_form',
+  FEEDBACK_VIEW = 'rating-label_feedbacks'
+}
