+++ src/enums/validationFunctionsEnum.ts
@@ -1,0 +1,6 @@
+export enum ValidationFunctions {
+  MIN_LENGTH = 'MIN_LENGTH',
+  PHONE = 'PHONE',
+  NAME = 'NAME',
+  EMAIL = 'EMAIL'
+}
